+++ jest.config.js
@@ -1,5 +1,24 @@
+const { run } = require('node:test');
+
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   testMatch: ['**/*.test.ts'],
+  // Separate test configurations
+  projects: [
+    {
+      displayName: 'unit',
+      preset: 'ts-jest',
+      testEnvironment: 'node',
+      testMatch: ['<rootDir>/src/**/*.test.ts'],
+      testTimeout: 10000, // 10 seconds for unit tests
+    },
+    {
+      displayName: 'integration',
+      preset: 'ts-jest',
+      testEnvironment: 'node',
+      testMatch: ['<rootDir>/integration-tests/**/*.integration.test.ts'],
+      testTimeout: 30000, // 30 seconds for integration tests
+    },
+  ],
 };